requirements.txt
streamlit==1.28.1
pillow==10.0.1
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python-headless==4.8.1.78
gpt4all==2.0.2
chromadb==0.4.18
llama-index==0.9.8
llama-cpp-python==0.2.11
sentence-transformers==2.2.2
transformers==4.35.0
pandas==2.1.1
numpy==1.24.3
python-dateutil==2.8.2 